Use alternative method to check for git updates
The current usage of 'LC_ALL=C git' to set the git locale does not work in some settings,
for example in the main virtual machine, distributed by chitanka.

The reported error is:

    sh: 1: LC_ALL=C git: not found

Is is also unknown if the current command will work in Windows.

For this reason, use an alternative method to check for git updates.